Add tests for new implementation of sort and create a mock js for tests
